Added model training on TPU. It fails on some datasets (perhaps if there are less than 2000 images?).
Changed dataset production to use a cloud function to partition the data because it takes longer than 30 seconds.

Changed dataset zip to use a cloud function to partition the records and one cloud function per partition to create a zip file for that partition.

Changed the UI to use tabs for Videos, Datasets, and Models.