Show training time for models.
Each team gets 120 minutes of training. When a training job is started, the user
specifies the max running time and that amount is subtracted from the team's
balance. When the job finishes (or is canceled or fails) the difference between
the max running time and the actual running time is added back to the team's
balance.