Terraform modifications
Removes bootstrap directory in favor of just using gsutil to create
the backend bucket for state storage

Move the gcs backend block to the root module so it will actually
get applied properly.

Fix the app engine resource so that it doesn't think there are changes
on each and every apply.

Add missing debug parameters