Change all calls to datetime.now() to datetime.now(timezone.utc).
Create the action_entity when the action is triggered. Save the action_parameters in the action_entity.

Removed debugging entrypoints /performActionGAE, /performActionGCF, and /test.
Removed test.html template.

Added /doesVideoEntityExist. Use it when a video is being uploaded.

Allow user to delete a video before frame extraction is finished.

In the Produce Dataset dialog, don't let the user enter out-of-bounds values for training and eval percents.

In the Start Training dialog, don't let the user enter an out-of-bounds value for the number of steps.

Added code for writing logs-based metrics for action, video, labeling, and tracking.

Added code to validate all request parameters.

Set response header "X-Frame-Options" to "SAMEORIGIN".

Upgraded the runtime version for training and eval jobs to 2.5.

Fixed bug in download dataset where sometimes it downloads multiple copies of each partition.

In models tab and monitor training page:
use "STOPPING" and "STOPPED" for when a training job is cancelled because it exceeded the maximum time.
use "STOPPING" and "STOPPED" for when an eval job is cancelled because it has finished all the evaluations.

Rearranged the Start Training and the Train More dialog boxes.