When validating createTimeMs parameters, allow the difference between user and server clocks to be up to 3 minutes.
Don't allow user to upload a video larger than 100 MB or longer than 2 minutes.

If frame extraction fails save an error message in the video entity and display it in a tooltip in the UI.
If the video has more than 1000 frames or resolution larger than 3840 x 2160, show that in a tooltip and don't extract frame.